# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Recova Backend API
  version: 0.2.0
  description: Kontrak API publik runtime Go Fiber Recova Backend.
servers:
  - url: /
    description: Base URL runtime service
paths:
...
  /api/v1/journals:
    get:
      tags:
        - journals
      operationId: getJournals
      summary: Mengambil daftar jurnal milik pengguna
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Daftar jurnal berhasil diambil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '401':
          description: Token akses tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags:
        - journals
      operationId: postJournals
      summary: Membuat entri jurnal pengguna
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JournalCreateRequest'
      responses:
        '201':
          description: Entri jurnal berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '401':
          description: Token akses tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Payload jurnal tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/community:
    get:
      tags:
        - community
      operationId: getCommunityPosts
      summary: Mengambil daftar postingan komunitas
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: category
          required: false
          schema:
            $ref: '#/components/schemas/CommunityCategory'
      responses:
        '200':
          description: Daftar postingan komunitas berhasil diambil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '401':
          description: Token akses tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Query kategori tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags:
        - community
      operationId: postCommunity
      summary: Membuat postingan komunitas
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunityCreatePostRequest'
      responses:
        '201':
          description: Postingan komunitas berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '401':
          description: Token akses tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Payload postingan tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Permintaan write community melebihi limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/education:
    get:
      tags:
        - education
      operationId: getEducation
      summary: Mengambil daftar konten edukasi aktif
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Konten edukasi berhasil diambil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '401':
          description: Token akses tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RefreshCookieAuth:
      type: apiKey
      in: cookie
      name: recova_refresh
  schemas:
    SuccessEnvelope:
      type: object
      required:
        - success
        - message
        - data
      properties:
        success:
          type: boolean
          const: true
        message:
          type: string
        data:
          type: object
          additionalProperties: true
        meta:
          oneOf:
            - type: object
              additionalProperties: true
            - type: 'null'
    ErrorEnvelope:
      type: object
      required:
        - success
        - message
        - data
        - error
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
        data:
          type: 'null'
        error:
          $ref: '#/components/schemas/ErrorPayload'
    ErrorPayload:
      type: object
      required:
        - code
        - requestId
      properties:
        code:
          type: string
        details:
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: object
                additionalProperties: true
            - type: 'null'
        requestId:
          type: string
...
    JournalCreateRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 5000
    CommunityCategory:
      type: string
      enum:
        - advice
        - motivation
        - story
        - question
        - assistance
    CommunityCreatePostRequest:
      type: object
      required:
        - content
        - category
      properties:
        title:
          type: string
          maxLength: 120
        content:
          type: string
          minLength: 10
          maxLength: 5000
        category:
          $ref: '#/components/schemas/CommunityCategory'
...
security: []